[Vitals] exclude supervisors and specialists from the old vitals query in US_IX (Recidiviz/recidiviz-data#35617)
## Description of the change

People are looking forward to this change and it'll make comparing
differences with the new queries easier, so I went ahead and implemented
the exclusion in the old queries.

I originally implemented it inside of
`supervision_officers_and_districts`, but found that it was
over-excluding some people who used to be TRANSITIONAL officers but
aren't anymore, so I opted to implement it inside of
`supervision_population_by_po_by_day`.

Sandbox at `dana_vitals_id_exclusions`. I looked at all the entities in
`vitals_summaries_differences` that were removed from "new" to make sure
they were a supervisor or transitional officer. For entities that are
still in both but changed values, they all were either locations in ID
(which makes sense why their values would change, since some officers
are now excluded that were in those locations) or were a location in ND
/ PO in IX who only changed in their 90d value (presumably because we're
now excluding someone from a date in the past).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes https://github.com/Recidiviz/recidiviz-data/issues/35104

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c19d925d0e73f394c367cdf6392dc82c6cec050c