[Insights] Modify supervisors querier logic to include supervisors who only have excluded officers (Recidiviz/recidiviz-data#35333)
Split this up from the other querier changes related to Recidiviz/recidiviz-data#33943 since it handles the supervisor errors we were seeing in prod. This specifically switches the query powering get_supervision_officer_supervisor_entities() to a left join on officer outlier statuses so that we don't lose all the supervisors who only have excluded officers.

See this slack thread for an example of this causing issues for supervisors attempting to log in to insights.

For testing, this PR adds a supervisor to the supervision_officer_supervisors.json fixture, and adds a single officer to the supervison_officer.json fixture for that supervisor. Does not add outlier status fixtures for the officer. In test_get_supervision_officer_supervisor_entities, the new supervisor is now pulled in.

GitOrigin-RevId: 8439448f9067b3b5c3781270ce51380613e049db