[Insights][Zero Grants] Modify querier and officer entity to send include_in_outcomes flag to FE (Recidiviz/recidiviz-data#35366)
Final PR for querier changes after the addition of a include_in_outcomes flag (and more officer rows) to the BQ officer product views.

The change modifies the SupervisionOfficerEntity to add an include_in_outcomes flag that can be consumed by the FE to distinguish between excluded and included officers. That flag will always be set to true, until a later PR when make the ExcludedOfficerEntity obsolete and start to send ALL officers (included and excluded) through the same endpoint.

This PR will be followed by a frontend PR that adds the includeInOutcomes field to the supervision officer Zod schema, and adds presenter and store logic to distinguish between excluded/included officers where necessary in the UI.

We can't send all officers through the same endpoint until the frontend is prepared to handle them.

GitOrigin-RevId: 4180541a8f89d625768b08d109242db86c90b972