[Justice Counts] Fix: Return `populations_dict` as empty dict if agency has no jurisdictions in Dashboard response (Recidiviz/recidiviz-data#35634)
## Description of the change

[This](https://recidiviz-inc.sentry.io/issues/6096383363/?project=4504532096516096&query=is%3Aunresolved%20issue.priority%3A%5Bhigh%2C%20medium%5D&referrer=issue-stream&statsPeriod=90d&stream_index=0)
Sentry error was surfaced because the dashboard API was visited for an
agency that has no jurisdictions. The
`AgencyJurisdcitions.get_agency_population` function returns a variable
called `populations_dict` that is only defined if the agency has
jurisdictions. This PR defines `populations_dict` regardless of the
agency's jurisdictions.

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 663f156c16d0d39247613c285d0e374f2b09738b