[Workflows] Fix minor type annotation issue in querier (Recidiviz/recidiviz-data#35761)
An issue arose last week where the `FullOpportunityInfo` argument to
`OpportunityConfigResponse.from_opportunity_and_config` was wrongly
annotated as `OpportunityInfo`, so the method appeared to work but
actually passed in bad arguments to the constructor. That bug was fixed
in Recidiviz/recidiviz-data#35511, but meanwhile, this PR changes the relevant type annotations,
and removes a test case where we pass `OpportunityInfo` in to
`OpportunityConfigResponse.from_opportunity_and_config` since we aren't
(and were never) doing that.

GitOrigin-RevId: 3e1d5a1e99b4cf61c28c1ed0f3b793b0abf42ee2