[Schema] Create StateEntityMixin and Document Normalization Created Fields (Recidiviz/recidiviz-data#35707)
## Description of the change

This change came about as I was working on a new version of schema
documentation generation. I initially wanted to tie descriptions to
`CachedAttributeInfo`, but found I really just wanted to pull
descriptions from the class. We could have also have added the
descriptions to a `metadata` dict in the class definitions, but then
we'd need to maintain that for both entities and normalized_entities. I
also found that approach a bit harder to read.

So, I made `StateEntityMixin` to house functionality that we want for
state and normalized_state entities. An immediate use for this besides
these field descriptions was the state_code field. Now future us will
only need to update the one field!

Next steps will be:
* Getting these descriptions when building BQ schemas
* Using BQ schemas for state and normalized state GitBook docs

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Supports Recidiviz/recidiviz-data#32861

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fdd15606e0f16543c5e79e4cc3b172f443364dad