[US_PA][Workflows] Add special conditions to case notes in PA opps (Recidiviz/recidiviz-data#35674)
## Description of the change

Add special conditions to case notes in PA opps so that they will
display in the sidebar for the agent to review. Per TT Nate Simon, only
including special conditions related to treatments/evaluations for now.

Because the format of special conditions is changing (Recidiviz/recidiviz-data#35570), this PR
also updates the "not serving on sex offense protocol" criteria which
pulls from special conditions.

Note - the new special conditions format is still not ingested, but we
want to get all these PRs merged this week to be ready for TT calls the
week of 12/16. So, I've created a sandbox where you can run the
opportunity query and get the actual results as they will appear once
the changes have gone through to normalized_state:
`recidiviz-staging.sophie_scratch.state_supervision_period_sandbox`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#33217

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b465cc22289f2172e0c75eb6782a891f1b67286e