add state code arg to cleanup_user_overrides script (Recidiviz/recidiviz-data#35780)
## Description of the change

This script exists to cleanup places where roster and user_override
match so that we can sync attributes of staff members that change over
time, even if they were once part of user_override. This is useful for
cleaning up after the first roster sync in a state, because a lot of
what the prep script does is write information into user_override. I
went to run it today (since we're pretty overdue for cleaning these out)
and noticed it doesn't let you filter by state, which made me feel a bit
uncomfortable, so I went ahead and added that.

Tested via unit tests and by running in staging against MI rosters.
Confirmed that it made the expected changes and did not seem to change
data for other states.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5dc12bb722812c6a6e22985d3dd0e02ca826248e