[Workflows] Apps Script Fix Statewide MAU and WAU (Recidiviz/recidiviz-data#35698)
## Description of the change

Changes how we calculate statewide supervision and facilities MAU and
WAU.

Previously, we were calculating statewide supervision and facilities MAU
and WAU by averaging the supervision and facilities MAU and WAU over all
workflows. This calculation is not actually what product wants to
display in the monthly leadership reports. Rather, product wanted us to
calculate the total number distinct of active users for any workflow
(de-duplicated) divided by the total number of registered users.

We accomplished this by created new aggregated metrics in Recidiviz/recidiviz-data#34896 and
then changing the calculations for `{{mau_pp}}`, `{{wau_pp}}`,
`{{mau_facility}}`, and `{{wau_facility}}` in the apps script code.

Slack thread related to this issue can be found
[here](https://recidiviz.slack.com/archives/C07CMGFR1PT/p1729002109359959?thread_ts=1728664718.488979&cid=C07CMGFR1PT).

#### How I Tested

Prior to this change, the statewide supervision and facilities MAU and
WAU for the Michigan report ending on 2024-12-01 looked as follows
(example report
[here](https://docs.google.com/document/d/1jLYG3gFCtmpz9jC0ee68PDqR4xAVbx1gF3vXJ8NlnS4/edit?tab=t.0#heading=h.7tap3soi8gte)):
![Screenshot 2024-12-05 at 12 09
57 PM](https://github.com/user-attachments/assets/0767464b-cbaf-4b51-8949-b40e64fba478)
![Screenshot 2024-12-05 at 12 10
03 PM](https://github.com/user-attachments/assets/0899acfe-84de-4331-babb-013f33860099)

With this change, the numbers now looks as follows (example report
[here](https://docs.google.com/document/d/15C_7VeAHJSL-2hTqC06-JTIbsyk22LmnOtdn4zXRwVw/edit?tab=t.0#heading=h.7tap3soi8gte)):
![Screenshot 2024-12-05 at 12 10
45 PM](https://github.com/user-attachments/assets/33872376-76f2-418c-a72b-dab56c96a604)
![Screenshot 2024-12-05 at 12 10
50 PM](https://github.com/user-attachments/assets/1dbb23e7-a342-4417-819c-c3cd752d783a)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes
https://app.zenhub.com/workspaces/polaris-614c9f27fa19b10015ddd495/issues/gh/recidiviz/recidiviz-dashboards/6663
Related to Recidiviz/recidiviz-data#34896

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@Michelles-MacBook-Pro.local>
GitOrigin-RevId: 10c0f26219fe45295508ee65c2b182e94a00f4f0